build: Removing tools/bazel wrapper script
This bazel wrapper script was added three years ago as a workaround
to an issue with the version of Bazel at that time. The issue was that
it was not possible out of the box to run multiple applications
concurrently in IntelliJ.

https://github.com/bazelbuild/bazel/issues/2337

Since at least July of 2018, this issue has been fixed and this
script can now be removed.

If it ain't broke, why fix it? Because in our case, it broke.
There have been situations where, whenever you make a change that causes
Java compilation to fail, IntelliJ will stop recognizing all Java files
in that package as sources. Very annoying.

At the time of writing, this phenomenon can be seen when editing any
file below tink-backend/src/agents-platform/.

The conditions to trigger this phenomenon are not exactly understood,
but I made a minimal working example to demonstrate it:

https://github.com/Sebelino/tink-backend/tree/wtf-unsynched-intellij-bazel-problem

1. Check out the branch
2. Import src/agents-platform/agents-platform.bazelproject into IntelliJ
3. Bazel sync
4. In AgentsPlatformFutures.java, change "public" to "ublic"
5. Bazel sync

Building will fail during sync (expected), but
AgentsPlatformFutures.java will no longer be recognized as a source file
by IntelliJ (unexpected).

Next:

1. In AgentsPlatformFutures.java, change "ublic" back to "public"
2. Bazel sync. The file will now be re-recognized as a source.
3. Remove the entire hoy/ directory
4. Bazel sync
5. In AgentsPlatformFutures.java, change "public" back to "ublic"
6. Bazel sync

Building will fail during sync (expected). However,
AgentsPlatformFutures.java will remain recognized as a source by
IntelliJ, albeit labeled "(unsynced)". This is the expected behavior
that I am used to.

I have no clue why the bug is triggered if there are two Bazel packages
as opposed to one.

But anyway, it seems that removing this obsolete script fixes the issue
altogether. I have done some testing and it seems the sources are now
always recognized, regardless of circumstances.