fix(supplementalrequester): inject requester
NewAgentTestContext was not a valid `SupplementalRequester` for
WireMock tets since it had a dependency on AgentTestServerClient.

This commit extracts SupplementalRequester functionality into two
separate components that can be injected depending on if we are
performing an agent test that should use agent test server, or a
WireMock test that should not.

Currently the MockSupplementalRequester only supports bankid. The
other types of supplemental information require more features to be
implemented.