fix(BEC): concurrency issue related to unsafe lazy loading technique
The previous version of algorithm looked like it is correct, but it
works with the assumption that the worst thing that can happen is to
have a null as a stale data - and it protects from it.
But the worst case is subtle, threads will not see stale reference to
the object but would see stale values for the object's state. Thanks to
instruction re-ordering threads would see partially constructed objects.

According to changes made to java memory model since JDK 5.0 volatile
keyword instructs JVM to not reorder instructions (sfence and mfence x86
instructions).

Aside from the fact that now lazy loading works properly, it is probably
not justified in any way to use it in our case and using synchronize on
the whole method is sufficient protection (slow uncontended
synchronization and slow JVM startup are not the case anymore).