feat(system-tests): Fetch image from Tink's private registry instead
quay.io has a pretty bad track record of incidents related to uptime:

https://status.quay.io/history

We therefore shouldn't depend on it being available in the CI pipeline.
The stability of our CI pipeline is the product of the individual
uptimes of the services it needs to fetch resources from.

gcr.io, on the other hand, has actual service-based SLAs and
historically much better uptime as well as faster time to resolution. I
am therefore reconfiguring our Testcontainers-based test to fetch the
ryuk image it depends on from our private gcr Docker registry instead.

https://cloud.google.com/storage/sla