fix(Bankia agent): fix error handler in login flow
Bankia will throw different codes during the authentication depending on the fault. Have added some of them to a mapper to check if it is related to wrong credentials. If not the agent will throw a loginerror with the message 'Unknown login error'

Bankia will also occasionally throw statusCode 428. This comes with no decription. Will catch it, and let the authentication process continue to see what will happen.