feat(Wiremock): Use package loader for decoupled mode
It makes more sense to use the package loader that we use for decoupled mode since Wiremock tests require to use the same modules.

Thanks to this change, we could get rid of extra module that is used for the Wiremock test for Barclays.

I'm not removing "withAgentModule" method though since it might be useful in the future to be able to override a custom module for decoupled mode during Wiremock tests. I want to keep this option available.