fix(LF OB): sessionHandler for LF OB
We have a lot of refresh failures. I believe that the cause of this is that LF only allows for a 20 minute period of fetching data after SCA. after that the consentId is invalid.

However, we store a token that can be refreshed much longer than 20 minutes. Our framework will in case of a valid access token think that we are good to go and skip the authentication even if the consentId has expired.

The result of this is that we can see a successful refresh call, but when we tries to get accounts we get SERVICE_BLOCKED.

Have implemented a SessionHandler that will check if our consentId is valid (by fetching accounts). If not, we remove the token from storage so that we can start a new Authentication process.