fix(AmexOB): fix logic in SessionHandler
the previous commit with the Amex SessionHandler did not work properly. It did not logout the user after a refresh. This commit is an attempt to fix this by having the logic in the keep-alive method.

Also, the SessionHandler did not check if the access token was expired or not. That could lead to us throwing away the token if its expired without first revoking or refreshing it.