fix(DanskeBank/OB): added filter for re-authenticate error
Part of CATS-758. DB will throw a 403 error with the message that the user needs to reauthenticate. This happen even if we successfully refreshes a token.
To handle this, I added a filter that specifically looks for it and if it happens, it removes the token from storage. This is done because the agent will otherwise try to use the token and if the refresh is ok, the agent will think that the token is valid.