fix(Amex OB): add Filter plus other fixes
Amex has fixed their poor Api by letting us renew a token. Thus, the 'special' fixes from previous commits are now reduntant and are removed by this commit.
Instead of a sessionhandler I added a filter that checks if the token is valid or not. If its revoked for some reason, I throw a CONSENT_REVOKE exception and remove the token from storage (this is needed to enforce the user to authenticate again).

I did a small refactor of the test, renamed agents to manual