feat(icabanken/nxgen): Truncate source message for payments
Icabanken does not support messages longer than 25 characters and it will lead to a 400 and a failed payment.
For now, this message will be truncated as we have customers that automatically populate the message from invoice data.
We can not cancel the payment with an end user message as the end user can not act on.