fix(Nordea/OB): update transaction description Nordea Fi
For transaction description in Nordea OB in all markets, we're setting the transaction description to narrative if it's not null or type_description otherwise.
For Finland, this results in useless descriptions, and a more useful one would be to use counterparty_name.

I will pass a market specific responseClass from the agent, to the client, that will determine how the transaction description will be set. Only Fi will use a specific one, while the other markets will use a common one.

I verified that it was working by running wireMock test for FI and SE.