fix(Amex OB): fix IllegalstateException during pagination
When paginating we are trying to fetch transactions for more than 90 days. Amex response with a 400 statusCode and a message. This will be catch and handled in the transactionFetcher.

The previous commit made the filter catch the error and throw a IllegalStateException. I fix this by checking the error message. If its due to pagination, I will let it slip through the filter