feat(WorkerContext): consolidate collected/cached account data
Currently we collect data from the agents in different maps in different contexts (one extends the next). However, we currently only filter two of these lists (sometimes) based on the whitelist requirements.

This commit consolidates all (I believe) data attached to an account so that it's easier to both keep track of and to uniformly filter/restrict.

The goal is to mimic/copy the behavior of the current code handling these different data (e.g. if they overwrite previously cached data or appends to it).