fix(LF OB): add missing header
Header TPP-NOK-REDIRECT-URI was missing during the authentication. This would cause a 404 error when cancelling bankID.

With the parameter I now got a callback when cancelling the authentication. The callback do not contain an 'error' parameter. Instead it has an empty parameter called pickup.

I temporary added a check that will look for the 'pickup' parameter and throw ThirdpartyApp.CANCEL.exception if it is present.