fix(AgentWorkerContext): Whitelist config is whitelist req
A whitelist configuration request will trigger a refresh but the context doesn't know it's a whitelist refresh (this commit corrects this!).

This caused an error message to be printed, `Account not found in updated Accounts list. This should not happen and might mean that Agent is not updating all Accounts separately.`, because accounts had been filtered by the whitelist.