feat(AWC): compare transaction caches
For debugging purposes only, this code will be removed when verified.

The goal is to compare the transactions that would be processed and sent to system. Therefore we use the filtered list of accountData as the old code uses the updatedAccounts list (only accounts that have been processed and sent to system).