fix(AgentTestContext): process accounts when cached
There was a bug in the agent test contexts where only accounts that had transactions would be processed. Since all of our logic is based on that an account has been processed I moved this code into the cacheAccount() method.

In the normal/production context we do this operation from a worker command.