fix(SDC): Make exceptionFilter null-safe
When providing pin length longer than 4, Sdc does not return X-SDC-ERROR-MESSAGE header,
hence there was added check of this header and Unit Tests of SdcExceptionFilter.
Add handling of invalid pin in a reponse body.
Also as a part of this commit, SdcConstants was unified.
In LoginError was added proper LocalizableKey for PIN exception.

ITE-1139 ITE-1153