fix(Nordea/RE): Using calculated dueDate
Make sure that we use dueDate that we calculated ourself in case the transfer's dueDate is null. Otherwise the payment in the bank will always have a dueDate whereas in cases where we don't get a dueDate fro, customer, we are comparing the existing payment's dueDate to "null" and that will always result in false. Thereby, duplicate payment.