build(system-tests): Created duplicate junit_test for easier local runs
The target named `service` will be fetching the ryuk image from the
public registry quay.io, while the target named `ci` will fetch it from
our private gcr.io registry as before.

The upside is that it will be easier for developers to run the system
tests locally, because there is no need to authenticate to our private
Docker registry before running the test.

The downside is that quay.io has lower availability than gcr.io. But
that is not so important for running the test locally. High availability
is important for CI builds, however, which is why they will continue to
fetch the image from gcr.io.