build(agents-framework): Adding second tink-backend git_repository rule
Having a second git_repository rule pointing tink-backend is a crutch to
implement versioning of the agents-platform agents-framework jar, which
is used in both repositories. We want the version of the jar used in
tink-backend-aggregation to vary independently of other reasons for
updating the revision of the tink_backend commit hash (e.g. changes to
providers in tink-backend). This enables Agent Platform to continue
making changes to tink-backend without blocking developers working in
with the framework tink-backend-aggregation in the short term.