fix(rate-limiter): Ensure only one RateLimiter is created per provider
Due to the `Provider` class not overriding `hashCode` or `equals`, our
RateLimiter would continually create new rate-limiter for ever request,
as each request would have a new Provider object. The default in Java
being identity checks for both `equality` and `hashCode`.

This evenutally leads to thread-exhaustion within the
`aggregation-service`, as a background refresh can be tens of thousands
of requests.