fix(lcl): make unit tests independent from current time
Current implementation of tests for LclAccessTokenProvider class is
based on results of OAuth2Token equals/hashCode methods. Those methods
rely on field which value carries current time in seconds and is
calculated with usage of System.currentTimeInMillis. It means that 2
tokens, created with same values (e.g. refreshtoken, accesstoken,
expiryinseconds) may be not equal when tested with equals/hashCode
methods.

As a result we got unit tests failing randomly.

#ITE-1503