test(lifeCycle): Verify `DropWizard` shutdown behaviour
Shows that an exception in `Managed::stop()` in an implemented class
will cause all subsequent `Managed::stop()` calls to get skipped. This
behaviour does not show which `Managed` service is causing this, only
the behaviour.

This also includes a bump of `tink-backend` version, to inlude the
`ManagedSafeStop` class as a drop-in replacement for `Managed`