chore(resource-limits): Lower resource usage of AS
This partially reverts commit a48aff410d1e2f02b761dbedae2c4bef773f199e,
as the root-casue of that incident was not memory usage but thread
leakage.

Keeping the larger number of replicas, but lowering the resources for
them should still allow the same performance as we have today.