chore(alerts/SE): Adding threshold range alert for login SR
The problem with current threshold setup is that an agent with login SR of 75% will trigger LoginSuccessRateBelow90WithAttemptLimitSE, LoginSuccessRateBelow80WithAttemptLimitSE in parallel. Expected would be to get paged for LoginSuccessRateBelow80WithAttemptLimitSE only. Attempting to fix this by adding threshold ranges. To not have duplicated queries added precomputations. Only added one alert alongside existing one to verify that it works as expected, will then proceed to exchange the existing alerts.