chore(logging): hash sensitive data
[MINI-57]

This PR changes the manner sensitive data are being masked in log file stored in s3.
Instead of removing sensitive data it is replaced with SHA-256 hash converted to Base64 and shortened to 4 signs.
Before:
   ***MASKED***
After:
  **HASHED:Bs64**

The purpose of this chore is to see if some values change during refreshes process or between them.

This PR does not change how Headers are being masked, only body.