fix(SparebankenSor): request activation code later
We should move sending request for activation code after we receive information about
reference words.

In current implementation there was a chance that request for activation code was send
before we receive confirmation for reference words. It occured when user confirmed BankID,
bank accept it and reference words were not confirmed.

In current solution we will not send request for activation code until we receive information
about reference words.

I know that it is somehow ugly - especially empty implementation, but it is working.
Bank ID native flow will solve this issue.

ITE-1266