feat(login): introducing session error enum types for consent issues
Before that commit, we had to throw BankSideError in case consent is invalid. This was causing credentials status to set temporary error and force a user to refresh again which makes the user experience worse. With this commit, we introduce consent related errors in SessionError type. Agents can throw them instead to retrigger authentication process without changing credentials status in case of a consent-related problem.

Agents that will throw these new errors must ensure that they clear consent in storage before and this task is left to market squads.

Also, currently for these new errors, we emit the same events (BankServiceError.X), this is to avoid a data migration at that point.