fix(hsbc): IFD-1722 fix refreshing the access token
In contrary to most banks, scope of the refresh token call must be openid accounts, not only accounts. Funny thing is, that this is the proper implementation, but is only followed but HSBC an FD - all the others do it wrong way. M&S is not affected (so in fact does it wrong), despite being in the same group as HSBC.