fix(ICA/RE): Update transaction fetching to match app
Before the app only sent toDate as a query param, but now uses fromDate and toDate. The fromDate in the transactions response now corresponds to the oldest transaction in that response, using that date minus one day to set the next toDate. Using this date instead of the oldest transactions date will allow us to continue pagination in the case where there are no transactions in the response. Max history in app is 17 months, so that's the max fromDate we'll use.