fix(Nordea): Change access token error handling
Never makes sense to return an error bankIdStatus after user has already signed. If no login code is present throw an IllegalstateException as that's an unknown error. Stop defaulting to bankIdStatus.NO_CLIENT for HttpResponseExceptions, it hides real errors. Created a new entity for login code response, it's confusing to refer to it as a bankid autostart response just because both contain a code.