fix(Nordea/SE): Use continuationKey for complete pagination
If users have a lot of transactions we get a continuationKey that we need to use in the next request. The app sets the startDate and endDate query parameters even though there's a continuationKey, therefore we do the same. We're not null checking the continuationKey as .queryParam() checks it and doesn't add the query param if the value is null.