log(context/processaccounts): new metric on number accounts per refresh
the existing metric we have were per account type, which will not show
the number of accounts returned by the entire refresh.

What triggers this PR to be added in the first place is to see how
often we return 0 (zero) accounts for an entire refresh in UK,
but I made the metric more generic than that to maybe be useful for
someone else in the future.