feat(NordeaNoPoc): Get bankID redirect URL from provider payload
ITE-1549
Redirect url always have to point to tink api. Normally, in the case of OB provider,
we can get this url from agent configuration. We can be sure it is set to some not empty
value, because it is in secrets that client had to upload in order for the provider to work.
However, this provider is RE, so we can not ask client to upload secrets for it. We could still
get some redirectUrl from "common configuration" (which is maintained per appId) if that client
has uplaoded any secrets for any OB provider for the app they're using. Of course, this assumption
is not always valid, e.g. if our client just made an account on Tink console and want to test
this particular provider without uploading any secrets for any other provider so far.