feat(dk-coop-password): BecAgent - use user's locale to initialize NemID authentication
ITE-1525
The NemID request's locale parameter is responsible for the language of notification
received on user's mobile phone. We have to change it to support English & Danish.

Unfortunately, changing this parameters also affects the language of API error responses by which
we categorize what went wrong. I've found 1 original translation for NO_MORTGAGE error by running local
integration test, but translations for the other 2 ones have to be figured out from production logs.
This means that agent would probably fail a few times in a following days and we need to monitor it and make a quick fix.