fix(DanskeBank RE, NO): handle bankID-app authentication method
[ITE-1829]
Previous method to check if it is mobile bank id was incorrect, because
it was returning true also for BankID-app in some cases.
If we recognised that it is mobile bank id, but in fact it was BankID
app then we were failing trying to get reference words.
If we did not recognised it as mobile bank id, then we were trying to
get list of auth methods and choose mentioned method. But in case of
users with BankID-app it resulted in two prompts - one from the app and
second as reference words.
If user accepted BankID-app fast enough, then we were failing because
we were trying to go back and choose other auth method, which was
impossible.

Handle mobile bank id first, then check if it is bankID-app (if yes,
then wait for user to accept it), then if it is other method - go to
auth methods list and choose mobile bank id.