fix(ICA/OB): Check customer status after refresh token is updated
Also moved setting of the token where it should be. We shouldn't throw a customer related error before the refresh token has been updated by the oauth2 controller. For users with KYC error we'll get a new refresh token. Since we threw the error before letting the controller update the refresh token the new token would not be stored, and next time user would get "invalid token" no matter if they fixed the KYC problem or not.