refactor: remove Managed interface
Removed DropWizard-specific Managed interface from AgentDataAvailabilityTrackerClient interface. The reason is, AgentDataAvailabilityTrackerClient interface will be used by the inner-client (which will be used both by AP and AS) mentioned in the previous commit and for this reason we want this inner-client to not depend on DropWizard. Since the SaAgentDataAvailabilityTrackerClientImpl extends ManagedSafeStop, we don't need the AgentDataAvailabilityTrackerClient to extend Managed.