fix: Fixed system tests
To ensure that we always bind AsAgentDataAvailabilityTrackerClient or AsAgentDataAvailabilityTrackerClientMockImpl when we have an object with AgentDataAvailabilityTrackerClient interface, I replaced all AgentDataAvailabilityTrackerClient with AsAgentDataAvailabilityTrackerClient. After I move the common module to tink-backend I can refactor the name, this is for fixing the system tests and to be on the safe side for now.