feat(DanskeBankDK): Throw ThirdPartyError instead of CREDENTIALS_VERIFICATION_ERROR when NemID is timed out/cancelled
ITE-1856
Currently when NemID 2FA polling times out or user rejects it in code app we're throwing CREDENTIALS_VERIFICATION_ERROR
which is too broad and have too general meaning "Your credentials could not be verified. Please try again.".
Moreover in DataStudio it is mapped to TINK/UNKNOWN_ERROR when it should be categorized as USER_ERROR.
To fix both issues we should throw a corresponding ThirdPartyError.