feat(DanskeBank): Add TimeoutFilter to handle bank's side connection issues
ITE-1856
In Kibana we see a lot of HttpClientExceptions with message "Connection reset". The reason for those errors
lies on a bank's side and we want to have them categorized as ASPSP_ERROR in DataStudio. To fix that we should
apply filter that will map http connection errors to BankServiceError.BANK_SIDE_FAILURE