feat(LF): handle issues when fetching OAuth2 token
LF has some issues with replying after request is sent. We are trying to retry that multiple times and when
finally they replies - we are receiving message that authorization is expired.

Because our implementation is correct, there is almost no chance that EXPIRED_AUTHORIZATION_CODE message will be returned
in the first attempt. It covers only retry situation.

TC-3929