fix(Belfius): Fixing basic error halding in Belfius RE
Also this is an entrypoint to error handling refactoring in Belfius.
Errors were consumed by validator inside api client throwing IllegalStateException almost everytime. This is not acceptable for AP framework.
This commit moves handling of most common errors to an external validator used in AP Step.
More work has to be done to completely remove validating in api client.

[MINI-576]