fix(N26): Adding AgentPlatformHttpResponseStatusHandler to the base class of AgentPlatformAgent
Without this change `DefaultResponseStatusHandler` is added to `TinkHttpClient`. 4xx statuses are treated like an error and it results with `HttpResponseException` exception being thrown. When using AgentPlatform api calls we want to be able to get all information in `ResponseEntity`. Thanks to that we will be able to easily manage all statuses in one place. By adding AgentPlatformHttpResponseStatusHandler to base class of AgentPlatformAgent all agents based on it will have this behaviour fixed