refactor(agent super classes): expose SuppInfoController
SupplementalRequester is a very lowlevel API that should not be used by
Agents directly. The lowest level of API (in regards to SI) is now
defined to be SupplementalInformationController. Helpers like
SupplementalInformationHelper and SupplementalInformtionFormer _can_ be
used, but they must use SupplementalInformationController if wanting to
interact with the user.

This commit does not yet restrict usage of SupplementalRequester, but
adds a FIXME comment. See this commit as a starting signal for migrating
away agents from using SupplementalRequester.
One exception has been made, and that is agents implementing
AbstractAgent instead of SubsequentGenerationAgent. Team owning these
agents should instead focus on migrating to use a more modern super
class (and with that using SupplementalInformationController over
SupplementalRequester).