fix(se/swedbank-ob): ignore bankid in auth request
TC-4091: Swedbank has made bankid field optional for authorize request in decoupled. For a user with single engagement e.g. Swedbank or one single branch of Savingsbank, it works as before. For multiple engagements, this enables us to provide a list to user to choose from and supply bankId later