fix(data-tracker): remove non-working code
We were trying to send transactions in two different ways (trying first approach first and if it fails trying the second approach) and we noticed that the first way does not work and the second way works. For this reason now we are removing the first way and directly execute the second way.