chore(alerts/SE): Ignore in SEB OB in refresh alert below 70
SEB OB frequently alerts due to issues we can't fix easily. If we go above 40 requests per day without PSU present we'll get an access exceeded error. SEB has been contacted about not having transaction description in the transaction list response. Main reason for this happening is that we have some customers that only refresh accounts during manual refresh. During BG refresh we refresh everything, so we will fetch all transactions in a BG refresh and hit the 40 request limit very quickly. The impact for end users in this scenario is none, since the customer is not interested in transactions. The other problem is that the certainDate that we use for limiting the transactions we fetch is set to a minimum of 45 days. With 4 refreshes a day we risk hitting the request limit TXN-303 has been created to investigate the possibility to shorten the certainDate. Since there is no quick fix for this that we can do, and user impact is very low, we now ignore SEB OB in the refresh alert for below 70% SR. If something goes really bad, we'll still be alerted on the below 60% alert.