fix(CBI/CredemAgent): handling invalid token
CredemAgent does not hit for a token because of a storage flag
CBI_MANUAL_AUTHENTICATION_IN_PROGRESS which was set to 'True' and
prevented OAuth token fetching from being called in
isAutoAuthenticationPossible().

In this commit token is fetched before evaluating the
possibility of performing auto-authentication. It would prevent
from making calls to Credem with an invalid token in storage (before
every executed step token would be checked for a validity).

A default state of the flag CBI_MANUAL_AUTHENTICATION_IN_PROGRESS when
there is no value for it in the storage is set to 'True' to avoid using
auto authentication with an invalid token.