feat(Xs2a): Adding protected methods in api client
When we want to create digest of the body we would have to override whole methods. By extracting creating request part we will be able to do it much cleaner and faster. Moreover final word was removed from Xs2aDevelopersTransactionalAgent.java to be able to add own implementation of those two classes in constructor