fix(Amex/OB): Remove foreign details model
Caused issues with parsing of the transaction model. The model seems to contain a lot of undocumented fields as well. Since we don't want to parse anything from it now and it's causing issues in production just parse foreign details as an object. If we need info from it later on we'll need to investigate what fields/formats it actually contains.