feat(api): add RefreshableSegment in RefreshRequest
A set of segments for which data are to be returned as a result of a refresh.
Will work similar to RefreshableItems but instead of dictating what will be fetched from a bank API it will instead limit/control what will be returned
(e.g. if BUSINESS is not specified in the set then any fetched business accounts and their transactions are removed from the result).