log(CrossKey): Tracing an issue with Spankki
Trying to determine if refresh token is always returned for the CrossKey provider. One of the clients has noticed that every time when doing a BR token is no longer valid. OAuth2AuthenticationController in autoAuthenticate method checks the refresh token and in my opinion it fails when trying to use a refresh token.  [MINI-849]