fix(bankid-controller): fail authentication if user is not available
The authenticate-auto case will just not work if it triggers SCA. User
won't get the SCA, which will make the agent timeout and fail, while
also blocking any further request (which is what the new agg
orchestration wants to do.

ideally the session expired exception should be thrown if we are here
no matter the request type. However, due to the lack of session handlers
in some agents, we do end up here in refresh-auto operations, which
we need to fix before rejecting completely. In those cases users will
have to double SCA instead, but it would still work.