feat(Buddybank): Fix BuddybankAgent
In this commit ClassCastException during cast UnicreditBaseApiClient to
BuddybankApiClient was fixed. In BuddybankApiClient a constructor which
as parameter takes UnicreditBaseApiClient was added.

In this commit Buddybank secrets error was fixed.

Credentials sessionExpiryDate was set based on a validation date from a
consent response.

BuddybankAuthenticator was changed to inherit from
UnicreditAuthenticator and use its methods.

[NZG-128][NZG-315][NZG-316][NZG-60]