refactor(legacy SEB): make java packages match on disk layout
This refactoring was applied automatically with a custom script.
The objective is to get the code-coverage measurement tools
working, as they require the on-disk layout to match the logical
layout defined in the java classes.

The Agent class was excluded from this refactoring, as it is
referred to by name outside the service