fix(DE/OB/various): Throw proper exception when weird selection
[NZG-311]

Some integrations with tink (klarna, mainly) can provide a weird value when presented with selection supplemental info.
They can provide a number that is way outside the range we expect, or a value that is not even a number we expect,
additionally, suppInfo mechanisms in TinkLink can sometimes provide empty map (by design), which also can lead to various exceptions here.

This fix, as ugly as it is, in multiple places, will throw proper exception when all this happens.