feat(bankid-controller): allow time properties for polling to be generic
Instead of always having a designated delay between each poll, this allows the user to set bankIdPollFrequency for each agent individually. If not stated, it will use the old value with 2 seconds between each poll.
The same logic applies for bankIdPollMaxAttempts. If not stated, it will be 90.
This also enables the user to set an initial delay, bankIdPollDelay, before the polling starts.