fix(dk/jyske): re-write account type mapper
when developing the agent it was believed that the regNo numeric field was related to the account type. When looking at production traffic we can see that is most likely not the case. So the only thing left that we have for deciding the account type is to check the account name. We know that opsparingskonto mean savings account.