chore(SE/Nordea/OB): Reduce requests in bankid flow
Nordea is really strict with rate limiting. Makeing requests related to bankID for 5 min leads to a lot of unnecessary requests. Added similar config as we did for Swedbank. Initial delay before polling of 2 sec, also reduced number of polls to 30. Set the initial delay and poll sleep to 2000ms since Nordea responds with "verify_after":2000 on the initiate/poll requests.