fix(FR/Bpce): Remove beneficiaries
Bpce claims that beneficiaries does not work on their end. "So, as this feature is not implemented  on our backends, this API method « GET /trustedBeneficiaries » (part of the STET swagger available on our 89C3 API Portal) will systematically generates errors." and they do not plan to add it in near future.