fix(Volksbank): Handle invalid request as SESSION_EXPIRED
We sometimes get this response even though we're using the correct token (needs to be investigated) or when we've failed to store the latest token in the db, and are trying to use an old expired token. We can't bring these credentials back to a good state since we don't have a valid token, user have to authenticate again.