fix(credentials-crypto): never consider cached data for one case
the case is when force authenticate flag is true.

The reason we introduce this is to not used any cached (_stale_) data
when trying to authenticate and set up a new consent. This is both a
conceptual change but will fix the bug we currently have where public
APIs refresh endpoint's forceAuthenticate flag set to true, that
invalidates the existing consent in DB, is not having any effect as we
will read from cache here in AggregationService