feat(HttpApiClient): Add detectors for storage variables
Detectors added for detecting token, clientId and consentId. The detectors will also populate the variable store upon detection. The detectors will act on insertion in persistent storage during runtime so that the http api client is updated with the most recent values. The detectors will also be used when creating the http api client, populating the variable store when reading from persistent storage.