feat(data-tracker): improvement on account event
When emitting DataTrackerEvent for accounts we take into consideration whether we tried to fetch transactions or not. If not so we send null value for "numberOfTransaction" field so in BigQuery we can distinguish cases where we tried to fetch transactions but we fetch 0 transactions and where we do not try to fetch transactions.